+++ plutus-executables/plutus-executables.cabal
@@ -23,6 +23,7 @@ source-repository head
 common lang
   default-language:   Haskell2010
   default-extensions:
+    BangPatterns
     DeriveFoldable
     DeriveFunctor
     DeriveGeneric
@@ -30,12 +31,20 @@ common lang
     DeriveTraversable
     DerivingStrategies
     DerivingVia
+    ExistentialQuantification
     ExplicitForAll
     FlexibleContexts
+    GADTs
     GeneralizedNewtypeDeriving
     ImportQualifiedPost
+    LambdaCase
+    MultiParamTypeClasses
+    NamedFieldPuns
+    OverloadedStrings
     ScopedTypeVariables
     StandaloneDeriving
+    TypeApplications
+    TypeSynonymInstances
 
   ghc-options:
     -Wall -Wnoncanonical-monad-instances -Wincomplete-uni-patterns
@@ -49,6 +58,17 @@ common lang
 common os-support
   if (impl(ghcjs) || os(windows))
     buildable: False
+
+library lib
+  import:          lang, os-support
+  hs-source-dirs:  src
+  exposed-modules: AgdaUnparse
+  build-depends:
+    , base               >=4.9   && <5
+    , bytestring
+    , plutus-core        ^>=1.36
+    , plutus-metatheory
+    , text
 
 executable pir
   import:         lang
@@ -88,7 +108,6 @@ executable uplc
     , Agda                             ==2.6.4
     , base                             >=4.9   && <5
     , bytestring
-    , containers
     , criterion
     , deepseq
     , filepath
@@ -98,6 +117,7 @@ executable uplc
     , optparse-applicative
     , plutus-core                      ^>=1.36
     , plutus-core:plutus-core-execlib
+    , plutus-executables:lib
     , plutus-metatheory
     , prettyprinter
     , split
