+++ plutus-metatheory/src/VerifiedCompilation.lagda.md
@@ -140,9 +140,7 @@ isTrace? ((tag , x₁ , x₂) ∷ xs) with isTrace? xs
 
 ## The certification function
 
-The `runCertifier` function is the top-level function which can be called by the compiler through the foreign function interface.
-It represents the "impure top layer" which receives the list of ASTs produced by the compiler and writes the certificate
-generated by the `certifier` function to disk. 
+The `runCertifier` function is the top-level function which can be called by the compiler.
 
 ```
